fix: tokens sorted correctly now and some code formatting changes
