+++ src/components/tokens/MyTradesTable/MyTradesTable.tsx
@@ -1,18 +1,6 @@
 import classNames from 'classnames'
-import {
-  useState,
-  useCallback,
-  useRef,
-  MutableRefObject,
-  useEffect,
-} from 'react'
+import { useCallback, useRef, MutableRefObject } from 'react'
 import { IdeaTokenTrade } from 'store/ideaMarketsStore'
-import {
-  bigNumberTenPow18,
-  calculateIdeaTokenDaiValue,
-  web3BNToFloatString,
-} from 'utils'
-import { sortNumberByOrder, sortStringByOrder } from '../utils'
 import { MyTradesRow, MyTradesRowSkeleton } from './components'
 import headers from './headers'
 
@@ -20,22 +8,22 @@ type MyTradesTableProps = {
   rawPairs: IdeaTokenTrade[]
   isPairsDataLoading: boolean
   canFetchMore: boolean
+  orderDirection: string
+  orderBy: string
   fetchMore: () => any
+  headerClicked: (value: string) => void
 }
 
 export default function MyTradesTable({
   rawPairs,
   isPairsDataLoading,
   canFetchMore,
+  orderDirection,
+  orderBy,
   fetchMore,
+  headerClicked,
 }: MyTradesTableProps) {
   const TOKENS_PER_PAGE = 10
-
-  const [currentHeader, setCurrentHeader] = useState('date')
-  const [orderBy, setOrderBy] = useState('date')
-  const [orderDirection, setOrderDirection] = useState('desc')
-
-  const [pairs, setPairs]: [IdeaTokenTrade[], any] = useState([])
 
   const observer: MutableRefObject<any> = useRef()
   const lastElementRef = useCallback(
@@ -52,127 +40,6 @@ export default function MyTradesTable({
     },
     [canFetchMore, fetchMore]
   )
-
-  useEffect(() => {
-    if (!rawPairs) {
-      setPairs([])
-      return
-    }
-
-    let sorted
-    const strCmpFunc = sortStringByOrder(orderDirection)
-    const numCmpFunc = sortNumberByOrder(orderDirection)
-
-    if (orderBy === 'name') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        return strCmpFunc(lhs.token.name, rhs.token.name)
-      })
-    } else if (orderBy === 'type') {
-      sorted = rawPairs.sort((lhs: any, rhs: any) => {
-        return numCmpFunc(lhs.isBuy, rhs.isBuy)
-      })
-    } else if (orderBy === 'amount') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        return numCmpFunc(lhs.ideaTokenAmount, rhs.ideaTokenAmount)
-      })
-    } else if (orderBy === 'purchaseValue') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        return numCmpFunc(lhs.daiAmount, rhs.daiAmount)
-      })
-    } else if (orderBy === 'currentValue') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        const tokenSupplyLeft = lhs?.isBuy
-          ? lhs?.token.rawSupply
-          : lhs?.token.rawSupply.add(lhs?.rawIdeaTokenAmount)
-        const ideaTokenValueLeft = parseFloat(
-          web3BNToFloatString(
-            calculateIdeaTokenDaiValue(
-              tokenSupplyLeft,
-              lhs?.market,
-              lhs?.rawIdeaTokenAmount
-            ),
-            bigNumberTenPow18,
-            2
-          )
-        )
-        const tokenSupplyRight = rhs?.isBuy
-          ? rhs?.token.rawSupply
-          : rhs?.token.rawSupply.add(rhs?.rawIdeaTokenAmount)
-        const ideaTokenValueRight = parseFloat(
-          web3BNToFloatString(
-            calculateIdeaTokenDaiValue(
-              tokenSupplyRight,
-              rhs?.market,
-              rhs?.rawIdeaTokenAmount
-            ),
-            bigNumberTenPow18,
-            2
-          )
-        )
-        return numCmpFunc(ideaTokenValueLeft, ideaTokenValueRight)
-      })
-    } else if (orderBy === 'pnl') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        const tokenSupplyLeft = lhs?.isBuy
-          ? lhs?.token.rawSupply
-          : lhs?.token.rawSupply.add(lhs?.rawIdeaTokenAmount)
-        const ideaTokenValueLeft = parseFloat(
-          web3BNToFloatString(
-            calculateIdeaTokenDaiValue(
-              tokenSupplyLeft,
-              lhs?.market,
-              lhs?.rawIdeaTokenAmount
-            ),
-            bigNumberTenPow18,
-            2
-          )
-        )
-        const pnlNumberLeft = ideaTokenValueLeft - lhs.daiAmount
-        const pnlPercentageLeft = (pnlNumberLeft / lhs.daiAmount) * 100
-
-        const tokenSupplyRight = rhs?.isBuy
-          ? rhs?.token.rawSupply
-          : rhs?.token.rawSupply.add(rhs?.rawIdeaTokenAmount)
-        const ideaTokenValueRight = parseFloat(
-          web3BNToFloatString(
-            calculateIdeaTokenDaiValue(
-              tokenSupplyRight,
-              rhs?.market,
-              rhs?.rawIdeaTokenAmount
-            ),
-            bigNumberTenPow18,
-            2
-          )
-        )
-        const pnlNumberRight = ideaTokenValueRight - rhs.daiAmount
-        const pnlPercentageRight = (pnlNumberRight / rhs.daiAmount) * 100
-
-        return numCmpFunc(pnlPercentageLeft, pnlPercentageRight)
-      })
-    } else if (orderBy === 'date') {
-      sorted = rawPairs.sort((lhs, rhs) => {
-        return numCmpFunc(lhs.timestamp, rhs.timestamp)
-      })
-    } else {
-      sorted = rawPairs
-    }
-
-    setPairs(sorted)
-  }, [rawPairs, orderBy, orderDirection, TOKENS_PER_PAGE])
-
-  function headerClicked(headerValue: string) {
-    if (currentHeader === headerValue) {
-      if (orderDirection === 'asc') {
-        setOrderDirection('desc')
-      } else {
-        setOrderDirection('asc')
-      }
-    } else {
-      setCurrentHeader(headerValue)
-      setOrderBy(headerValue)
-      setOrderDirection('desc')
-    }
-  }
 
   return (
     <div className="flex flex-col">
@@ -198,11 +65,11 @@ export default function MyTradesTable({
                       <div className="flex">
                         {header.sortable && (
                           <>
-                            {currentHeader === header.value &&
+                            {orderBy === header.value &&
                               orderDirection === 'asc' && (
                                 <span className="mr-1">&#x25B2;</span>
                               )}
-                            {currentHeader === header.value &&
+                            {orderBy === header.value &&
                               orderDirection === 'desc' && (
                                 <span className="mr-1">&#x25bc;</span>
                               )}
@@ -215,7 +82,7 @@ export default function MyTradesTable({
                 </tr>
               </thead>
               <tbody className="bg-white divide-y divide-gray-200 dark:bg-gray-700">
-                {pairs.map((pair: IdeaTokenTrade, index) => (
+                {rawPairs.map((pair: IdeaTokenTrade, index) => (
                   <MyTradesRow
                     key={index}
                     token={pair.token}
@@ -225,7 +92,7 @@ export default function MyTradesTable({
                     timestamp={pair.timestamp}
                     rawIdeaTokenAmount={pair.rawIdeaTokenAmount}
                     lastElementRef={
-                      pairs.length === index + 1 ? lastElementRef : null
+                      rawPairs.length === index + 1 ? lastElementRef : null
                     }
                   />
                 ))}
